+++ npbloom-core/src/commonMain/kotlin/space/yuvalinguist/npbloom/ui/SettingsState.kt
@@ -24,9 +24,16 @@ class SetLiveStringWidth(val liveStringWidth: Boolean) : SettingsAction {
 }
 
 @JsExport
+class SetPrettyNodeLabels(val prettyNodeLabels: Boolean) : SettingsAction {
+    override val key = "prettyNodeLabels"
+    override val value = prettyNodeLabels.toString()
+}
+
+@JsExport
 data class SettingsState(
     val autoFormatSubscript: Boolean,
     val liveStringWidth: Boolean,
+    val prettyNodeLabels: Boolean,
 )
 
 @JsExport
@@ -34,4 +41,5 @@ fun settingsReducer(state: SettingsState, action: SettingsAction): SettingsState
     when (action) {
         is SetAutoFormatSubscript -> state.copy(autoFormatSubscript = action.autoFormatSubscript)
         is SetLiveStringWidth -> state.copy(liveStringWidth = action.liveStringWidth)
+        is SetPrettyNodeLabels -> state.copy(prettyNodeLabels = action.prettyNodeLabels)
     }