+++ npbloom-web/src/currentVersion.tsx
@@ -1,4 +1,5 @@
-export const currentVersion: string = '0.7.1';
+export const currentVersion: string = '0.7.4';
 
 export const changesFromPreviousVersion: (string | JSX.Element)[] = [
+  <>Fixed bug that caused some clicks to randomly not count on certain operating systems.</>,
 ];