+++ HISTORY.md
@@ -1,4 +1,13 @@
 ## Alpha
+### 0.7.4 (2025-10-11)
+* Fixed bug that caused some clicks to randomly not count on certain operating systems.
+### 0.7.3 (2025-02-16)
+* Added simple strikethrough functionality.
+  Select any part of a sentence and click the Strikethrough button on the toolbox to draw a line through it.
+  (tnx Amit Benalal for requesting)
+### 0.7.2 (2024-11-11)
+* Added pretty rendering for bar notation and subscripts in node labels.
+  On by default, can be turned off via Settings. (tnx Amit Benalal for requesting)
 ### 0.7.1 (2024-11-10)
 * Fixed bug that caused corresponding nodes in copied and pasted trees to be linked.
 ### 0.7 (2024-11-03)